% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnf-tools.R
\name{is_cnf}
\alias{is_cnf}
\title{Determine if the expression (in a string) is in conjunctive normal form}
\usage{
is_cnf(s)
}
\arguments{
\item{s}{single string containing an expression}
}
\value{
Logical value. TRUE if expression is a CNF. Return FALSE if 
        expression is a valid boolean formula but is not in CNF.  
        Otherwise an error is raised.
}
\description{
Determine if the expression (in a string) is in conjunctive normal form
}
\examples{
# Expect TRUE for these CNF expressions
is_cnf("a | !b") 
is_cnf("(a | !b) & (c | d | e)") 

# Expect FALSE for these expressions
is_cnf("!!a")
is_cnf("!(a & b)")
is_cnf("!(a | b)")
is_cnf("a & (b | (d & e))")
}
