% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimacs.R
\name{print.sat_dimacs}
\alias{print.sat_dimacs}
\title{Print a DIMACs representation returned from \code{\link{literals_to_dimacs}()}}
\usage{
\method{print}{sat_dimacs}(x, ...)
}
\arguments{
\item{x}{\code{sat_dimacs} object}

\item{...}{ignored}
}
\value{
None
}
\description{
Print a DIMACs representation returned from \code{\link{literals_to_dimacs}()}
}
\examples{
literals <- c(1, 2, 3, 0, -2, 4, 0, 5, 6, 7, 0)
dimacs <- literals_to_dimacs(literals)
dimacs
}
