% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-core.R
\name{print.sat_prob}
\alias{print.sat_prob}
\title{Print 'sat' object}
\usage{
\method{print}{sat_prob}(x, ...)
}
\arguments{
\item{x}{'sat' object created by \code{\link{sat_new}()}}

\item{...}{ignored}
}
\value{
Invisible copy of original \code{x}
}
\description{
Print 'sat' object
}
\examples{
sat <- sat_new()
print(sat)
}
