% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-core.R
\name{sat_add_literals}
\alias{sat_add_literals}
\title{Add integer literals to a SAT problem}
\usage{
sat_add_literals(sat, literals)
}
\arguments{
\item{sat}{\code{sat} object as created by \code{\link{sat_new}()}}

\item{literals}{integer vector of positive and negative literal values.
Clauses should be separated using zero (0).  If the vector is not
terminated by a zero (0), then a zero will be added.}
}
\value{
None
}
\description{
Note: Adding clauses to a SAT problem can only be via adding integer literals
(via \code{sat_add_literals()}) or adding named expressions (via 
\code{sat_add_exprs()}) .  The two 
methods cannot be mixed within a single SAT problem.
}
\examples{
# Define a small SAT problem
sat <- sat_new()
sat_add_literals(sat, c(-1, 2))
sat_add_literals(sat, c(-2, 3))
sat_add_literals(sat, c(-3, 1))
sat_solve_naive(sat)


# Equivalent problem
sat <- sat_new()
sat_add_literals(sat, c(-1, 2, 0, -2, 3, 0, -3, 1, 0))
sat_solve_naive(sat)
}
\seealso{
Other ways of adding clauses: 
\code{\link{sat_add_exprs}()}
}
\concept{ways of adding clauses}
