% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-core.R
\name{sat_block_solution}
\alias{sat_block_solution}
\title{Block a given result - use this for manually finding multiple solutions
to a given problem.}
\usage{
sat_block_solution(sat, literals, verbosity = 0L)
}
\arguments{
\item{sat}{\code{sat} object as created by \code{\link{sat_new}()}}

\item{literals}{A solution. Vector of integer literals.}

\item{verbosity}{verbosity}
}
\value{
None
}
\description{
Block a given result - use this for manually finding multiple solutions
to a given problem.
}
\examples{
# Setup a problem and solve
sat <- sat_new()
sat_add_exprs(sat, "!a | b")
sat_add_exprs(sat, "!b | c")
sat_add_exprs(sat, "!c | a")
sat_solve_naive(sat)

# Block the solution where all values are TRUE
soln1 <- "a & b & c"

# Block this solution from occurring
sat_block_solution(sat, soln1)

# Now only a single solution returned
sat_solve_naive(sat)
}
