% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-cardinality-constraints.R
\name{sat_card_exactly_k}
\alias{sat_card_exactly_k}
\alias{sat_card_atleast_k}
\alias{sat_card_atmost_k}
\title{Add to SAT problem the cardinality constraint that "exactly k", "at most k"
or "at least k" of the given variables are true}
\usage{
sat_card_exactly_k(sat, nms, k)

sat_card_atleast_k(sat, nms, k)

sat_card_atmost_k(sat, nms, k)
}
\arguments{
\item{sat}{\code{sat} object as created by \code{\link{sat_new}()}}

\item{nms}{character vector of variable names}

\item{k}{k limit}
}
\value{
a character vector of expression strings representing the cardinality constraint
        in conjunctive normal form
}
\description{
This constraint is an implementation of Sinz's LTseq formulation.
}
\examples{
sat <- sat_new()
sat_card_exactly_k(sat, c('a', 'b', 'c', 'd'), k = 2)
sat$exprs

sat <- sat_new()
sat_card_atleast_k(sat, c('a', 'b', 'c', 'd'), k = 2)
sat$exprs

sat <- sat_new()
sat_card_atmost_k(sat, c('a', 'b', 'c', 'd'), k = 2)
sat$exprs
}
\seealso{
Other SAT cardinality constraints: 
\code{\link{sat_card_exactly_one}()}
}
\concept{SAT cardinality constraints}
