% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-core.R
\name{sat_literals_to_lgl}
\alias{sat_literals_to_lgl}
\title{Convert a sequence of integer literals to named logical vector}
\usage{
sat_literals_to_lgl(sat, literals, keep_dummy = FALSE)
}
\arguments{
\item{sat}{\code{sat} object as created by \code{\link{sat_new}()}}

\item{literals}{literals}

\item{keep_dummy}{Should dummy variables be kept?  Default: FALSE.
Dummy variables are those variables added internally (e.g. when adding
cardinatlity constraints).  Dummy variables all start with the 
prefix "dummy".}
}
\value{
Named logical vector indicating the names and boolean values of the 
        given literals
}
\description{
This function is used when a set of literals (Usually corresponding to a solution)
needs to be turned into a logical statement with named variables.
}
\examples{
sat <- sat_new()
sat_add_exprs(sat, c("(a | !b) & (!c | !d)"))
sat$names
sat$named_literals
sat_literals_to_lgl(sat, c(1, 3, -4))
}
