% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-core.R
\name{sat_new}
\alias{sat_new}
\title{Initialise a new SAT problem}
\usage{
sat_new()
}
\value{
\code{sat} object
}
\description{
Initialise a new SAT problem
}
\examples{
sat <- sat_new()
sat
sat_add_literals(sat, c(1, 2))
sat_add_literals(sat, c(2, -3))
sat
sat$names
sat$named_literals
sat$exprs
sat$dimacs
sat_solve_naive(sat)
}
