% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satire-package.R
\docType{package}
\name{satire-package}
\alias{satire}
\alias{satire-package}
\title{satire: Define and Solve Boolean Satisfiability ('SAT') Problems}
\description{
Boolean satisifiability ('SAT') tests whether there are sets of variables which satisfy a Boolean formula. This package provides tools for defining 'SAT' problems and includes a 'DPLL'-based 'SAT' solver based upon Davis & Putnam (1960) \doi{10.1145/321033.321034}. Built-in capabilities for Tseytin transformation (Tseytin (1968) \doi{10.1007/978-3-642-81955-1_28}) allows for the input of complex Boolean formula. Cardinalility constraints are implemented using commander variables (Klieber & Kwon (2007) \url{https://www.cs.cmu.edu/~wklieber/papers/2007_efficient-cnf-encoding-for-selecting-1.pdf}) and sequential counters (Sinz (2005) \doi{10.1007/11564751_73}). Tools for reading and writing 'DIMACS' files are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/satire}
  \item Report bugs at \url{https://github.com/coolbutuseless/satire/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
