% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseytin-transform.R
\name{tseytin_transform}
\alias{tseytin_transform}
\title{Convert a boolean formula to CNF using Tseytin's transformation}
\usage{
tseytin_transform(expr, dummy_idx = 1L)
}
\arguments{
\item{expr}{expression (as string or language object)}

\item{dummy_idx}{value to use for numbering the dummy variables.}
}
\value{
character vector of expressions in conjunctive normal form
}
\description{
Convert a boolean formula to CNF using Tseytin's transformation
}
\examples{
tseytin_transform("a | (b & c)")
tseytin_transform("!(a | b)")
}
