% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat-core.R
\name{sat_literals_to_lgl}
\alias{sat_literals_to_lgl}
\title{Convert a sequence of integer literals to named logical vector}
\usage{
sat_literals_to_lgl(sat, literals, remove = "^dummy")
}
\arguments{
\item{sat}{\code{sat} object as created by \code{\link{sat_new}()}}

\item{literals}{literals}

\item{remove}{regular expression for variables to remove when blocking solutions
and assembling values to return. Default: "^dummy" will block all
variables starting with the word "dummy" (as this is how the 'satire' 
package automatically creates dummy variables.)
If NULL no variables will be removed.}
}
\value{
Named logical vector indicating the names and boolean values of the 
        given literals
}
\description{
This function is used when a set of literals (Usually corresponding to a solution)
needs to be turned into a logical statement with named variables.
}
\examples{
sat <- sat_new()
sat_add_exprs(sat, c("(a | !b) & (!c | !d)"))
sat$names
sat$named_literals
sat_literals_to_lgl(sat, c(1, 3, -4))
}
